rpi5/voicemail-assistant/requirements.txt
pipecat-ai[daily,google]==0.0.68

loguru>=0.7
